% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTaetiology}
\alias{BTaetiology}
\title{BTaetiology}
\format{
A data frame with 225 rows and 24 variables:
\describe{
\item{pid}{Study identifier}
\item{hivstatus}{HIV status one of (Non reactive, Reactive, Unknown)}
\item{mtb bsi}{M. tuberculosis cultured from blood? (1 = yes, 0 = no, NA = no valid result)}
\item{malaria}{Malaria? (1 = yes, 0 = no, NA = no valid result)}
\item{uLAM}{Positive uLAM result?  (1 = yes, 0 = no, NA = no valid result)}
\item{Xpert}{Positive sputum Xpert? (1 = yes, 0 = no, NA = no valid result)}
\item{dengue}{Dengue diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{chik}{Chikungunya diagnosis as per case definition? Dengue diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{bact_target_pcr}{Pathogenic bacterial target found on PCR array card?  (1 = yes, 0 = no, NA = no valid result)}
\item{lepto}{Leptospirosis diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{SF}{Spotted fever group Rickettsiosis diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{ET}{Epidemic Typhus diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{borrelia}{Borreliosis diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{rift_valley_fever}{Rift Valley fever diagnosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{CrAg_LFA}{Positive cryptococcal antigen lateral flow assay on CSF (1 = yes, 0 = no, NA = no valid result)}
\item{bsi.bacterial}{Bacterial bloodstream infection as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{bsi.fungal}{Fungal bloodstream infection as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{csf.fungal}{Fungal meningitis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{csf.bacterial}{Bacterial meningitis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{inv.fungal}{Any invasive fungal infection as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{inv.bacterial}{Any invasive bacterial infection as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{tb}{Tuberculosis as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{arbovirus}{Any arboviral infection as per case definition? (1 = yes, 0 = no, NA = no valid result)}
\item{ricks}{Any rickettsial infection as per case definition? (1 = yes, 0 = no, NA = no valid result)}
}
}
\usage{
BTaetiology
}
\description{
Aetiology of sepsis of 225 adults with sepsis in Blantyre, Malawi
}
\keyword{datasets}
