% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTarraycard}
\alias{BTarraycard}
\title{BTarraycard}
\format{
A data frame with 225 rows and 12 variables:
\describe{
\item{pid}{Study identifier}
\item{array_card_xxx}{Was the target xxx identfied on array card testing? (TRUE = yes, FALSE = no, NA = no valid result)}
}
}
\usage{
BTarraycard
}
\description{
Array card multiplex PCR results for 225 adults with sepsis in Blantyre, Malawi
}
\keyword{datasets}
