% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTbc}
\alias{BTbc}
\title{BTbc}
\format{
A data frame with 225 rows and 22 variables:
\describe{
\item{pid}{Study identifier}
\item{lims_fe}{Sample identifier}
\item{bcult_anygrowth}{Any growth on aerobic blood culture? (1 = yes, 0 = no)}
\item{typhi}{Growth of Salmonella Typhi (1 = yes, 0 = no)}
\item{saen}{Growth of Salmonella Enteritides (1 = yes, 0 = no)}
\item{saty}{Growth of Salmonella Typhimurium (1 = yes, 0 = no)}
\item{esco}{Growth of E. coli (1 = yes, 0 = no)}
\item{klpn}{Growth of Klebsiella pneumoniae (1 = yes, 0 = no)}
\item{hib}{Growth of Haemophilus influenzae (1 = yes, 0 = no)}
\item{dip}{Growth of Diptheroids (1 = yes, 0 = no)}
\item{bacillus}{Growth of Bacillus spp. (1 = yes, 0 = no)}
\item{esco}{Growth of E. coli (1 = yes, 0 = no)}
\item{stpn}{Growth of Streptococcus pneumoniae (1 = yes, 0 = no)}
\item{crne}{Growth of Cryptococcus neoformans (1 = yes, 0 = no)}
\item{ahaemstrep}{Growth of Alpha-haemolytic Streptococci other than Pneumococcus (1 = yes, 0 = no)}
\item{acba}{Growth of Acinetobacter baumannii (1 = yes, 0 = no)}
\item{enfam}{Growth of Enterococcus faecium (1 = yes, 0 = no)}
\item{enfas}{Growth of Enterococcus faecalis (1 = yes, 0 = no)}
\item{micr}{Growth of Micrococcus spp. (1 = yes, 0 = no)}
\item{cons}{Growth of Coagulase-negative Staphylococcus (1 = yes, 0 = no)}
\item{stau}{Growth of Staphyloccus aureus (1 = yes, 0 = no)}
\item{prot}{Growth of Proteus mirabilis (1 = yes, 0 = no)}
\item{pathogen}{Any pathogen identfied? (1 = yes, 0 = no)}

}
}
\usage{
BTbc
}
\description{
Aerobic blood culture results for 225 adults with sepsis in Blantyre, Malawi
}
\keyword{datasets}
