% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTparticipants}
\alias{BTparticipants}
\title{BTparticipants}
\format{
A data frame with 225 rows and 22 variables:
\describe{
\item{pid}{Study identifier}
\item{calc_age}{Age in years at enrollment}
\item{ptsex}{Gender one of (male,female)}
\item{hivstatus}{HIV status one of (Non reactive, Reactive)}
\item{cd4_absolute}{CD4 cell count in cells per microlitre}
\item{hivonart}{For HIV infected participants, whether currently taking antiretroviral therapy. One of (Yes, No)}
\item{art_time}{Number of months on antiretroviral therapy}
\item{hivcpt}{For HIV infected participants, whether currently taking co-trimoxazole preventative therapy. One of (Yes, No)}
\item{ever_tb}{Ever been treated for TB? One of (Yes, No)}
\item{tbongoing}{Currently being treated for TB? One of (Yes, No)}
\item{screentemp}{Temperature in degrees celsius at recruitment}
\item{t0hr}{Heart rate in beats per minute at recruitment}
\item{t0sbp}{Systolic blood pressure at recruitment in mmHg}
\item{t0dbp}{Diastolic blood pressure at recruitment in mmHg}
\item{gcs}{Glasgow coma score at recruitment}
\item{ustand}{Unable to stand at recruitment}
\item{days_unwell}{Length of current illness in days}
\item{haemoglobin}{Haemoglobin in g/dL at recrutment}
\item{platelets}{Platelet count in cells x 10^9/L at recruitment}
\item{wcc}{White cell count in cells x 10^9/L at recruitment}
\item{sodium}{Serum sodium in mmol/L at recruitment}
\item{potassium}{Serum potassium in mmol/L at recruitment}
\item{co2}{Serum bicarbonate in mmol/L at recruitment}
\item{urea}{Serum urea in mmol/L at recruitment}
\item{creatinine}{Serum creatinine in mmol/L at recruitment}
\item{lactate}{Serum lactate in mmol/L at recruitment}
\item{urea}{Serum urea in mmol/L at recruitment}
\item{dxx_death}{Death by xx days, 1 = died, 0 = survived}
\item{t}{Follow up time in days}
\item{died}{Did participant die (1) or was alive (0) at the end of follow up time}
}
}
\usage{
BTparticipants
}
\description{
225 adults with sepsis in Blantyre, Malawi
}
\details{
Demographics, vitals, investigations, and outcome for 225 adults with sepsis in Blantrye, Malawi. Missing data given by NA in all cases.
}
\keyword{datasets}
