% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTtreatment}
\alias{BTtreatment}
\title{BTtreatment}
\format{
A data frame with 224 rows and 10 variables:
\describe{
\item{pid}{Study identifier}
\item{timeto_ab}{Time from ED enrollment to first administration of antibacterial agent in hours}
\item{which_ab}{Name of antibacterial agent administered}
\item{timeto_antifungal}{Time from ED enrollment to first administration of antifungal agent in hours}
\item{which_antifungal}{Name of antifungal agent administered}
\item{timeto_antimalarial}{Time from ED enrollment to first administration of antimalarial agent in hours}
\item{which_antimalarial}{Name of antimalarial agent administered}
\item{timeto_antitb}{Time from ED enrollment to first administration of antituercular  agent in hours}
\item{which_antitb}{Name of antitubercular agent(s) administered}
\item{iv_fluid_6hr}{Volume of intravenous fluid administered in the 6 hours from ED enrollment }

}
}
\usage{
BTtreatment
}
\description{
Treatments administered to 225 adults with sepsis in Blantyre, Malawi
}
\keyword{datasets}
