% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_functions.R
\name{make_kable_rowgroup_string}
\alias{make_kable_rowgroup_string}
\title{Helper to group variables together in Kable tables}
\usage{
make_kable_rowgroup_string(data, varnm)
}
\arguments{
\item{data}{Table to group vars (usually output from pretty_tb_df)}

\item{varnm}{Variable to group by}
}
\value{
Named character vector: values are variable names and vector names are number of rows
}
\description{
Use this as input into kableExtra:pack_rows
}
