% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_functions.R
\name{pretty_tbl_df}
\alias{pretty_tbl_df}
\title{Create a pretty table from a dataframe to pass to Kable for rendering}
\usage{
pretty_tbl_df(
  df,
  vars_to_char = NULL,
  r = 0,
  vars_to_specify_rounding = NULL,
  confint = TRUE
)
}
\arguments{
\item{df}{dataframe to make into pretty table}

\item{vars_to_char}{character vector of names of numeric variables that you want to treat as factors}

\item{r}{numeric, number of digits to round to}

\item{vars_to_specify_rounding}{named numeric vector c("varname" = m) to hange default rounding for variable varname}

\item{confint}{boolean, do you want to report confidence intervals?}
}
\value{
A dataframe with three columns:
variable: the original column name
levels: the levels of a character variable (one row for each) level OR
the text "Median (IQR)" for numeric variables
value: the proportion of observations for that level OR median (IQR) for
numeric variables
}
\description{
Take a dataframe and make a new summary dataframe to pass to Kable for
pretty table rendering. The input is assumed to have variables in columns
with observations in rows.Numeric variables are by default summarised by median and IQR.
Character vectors are assumed to be factors and a summarised as a
proportion for each level.
}
\examples{
pretty_tbl_df(mtcars)
}
