% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_functions.R
\name{sp_dc}
\alias{sp_dc}
\title{Helper to reliably round number to specified decimal places}
\usage{
sp_dc(x, k)
}
\arguments{
\item{x}{Number to round}

\item{k}{Number of decimal places}
}
\value{
String
}
\description{
Helper to reliably round number to specified decimal places
}
